feat(logging): global methods for logging and telemetry
Logging can be enabled by setting the `debug` key in local storage to `nillion:*`. This adds `window.__NILLION.enableLogging(overwrite: boolean)` to make it easy to toggle logging in the TS client.

Because it is now easier to control logging this commit no longer enables logging by default so that the user may control it.

Finally, exposes `window.__NILLION.enableWasmLogging()` and `window.__NILLION.enableTelemetry(addr: string)`.