feat: use Effect as common return type
The effect library provides ergonomic error handling and ADTs; however, its API is complex and could be a blocker to those less familiar with this style. So, this commit aims to strike a balance so that we benefit from the robustness offered without exposing the complexity to end users. This is done by adopting the effect library internally and then presenting `type Result<S, E> = { ok: S, err: null } | { ok: null, err: E }` as the result of `NillionClient` operations.

For example:

```typescript
declare client: NillionClient
const result = await client.fetchValue("store-id-aaa", "foo")
if(result.err) {
   // something went wrong and the err is on result.err
}
if(result.ok) {
   // the operation succeeded the data is on result.ok
}
```

The result type could be enhanced a bit further, but for the moment its deliberatly kept simple.