feat: make client constructors sync to avoid top level awaits
Both the `NilVmClient` and the `NilChainPaymentClient` had async constructor dependencies (e.g. `NilVmClient` depends on `await wasmBundle.default()` and `NilChainPaymentClient` depends on the async OfflineSigner).

Async is problematic in contexts like React you construct a `NillionClient` and pass it to `<NillionProvider value={client} />` at the top of the application's component tree. This means we're blocking, require management of the loading/error/success outcomes, and users must deal with top-level await effects.

To avoid that this commit changes all constructors to be sync and defers async behaviour to `connect(args): Promise<boolean>` instance methods. This means these clients can be constructed in a non-blocking fashion whilst depends are loaded in the background.