feat: drop keplr.experimentalSuggestChain use keplr.enable(chain)
The former configures keplr while the later activates the chain if its present. We should avoid configuring keplr because there's more room for error; suggesting to activate a chain is inherently less risky.

In testing I haven't managed to see a change in behaviour but this is the recommended set of calls from [the keplr docs](https://docs.keplr.app/api/).